+++ test/browser/test.ml
@@ -1,0 +1,1 @@
+let () = Alcotest.run "browser" [ Test_browser.suite ]